Fix: publishedAt column becomes bigint
Note: after 2038, it will not be possible
to store unix timestamps as integer
anymore.

+++ create_db.sql
@@ -7,5 +7,5 @@ CREATE TABLE news(
     title VARCHAR(200),
     description TEXT,
     url VARCHAR(255),
-    publishedAt INT
+    publishedAt BIGINT
 );
